Earlier warning fix was incomplete
and actually broke the whole thing. Carry type change through up to base class.

+++ phidgets_api/include/phidgets_api/ik.h
@@ -15,10 +15,10 @@ class IK: public Phidget
 
     CPhidgetInterfaceKitHandle ik_handle_;
     virtual void sensorHandler(
-      int index,
+      size_t index,
       int sensorValue);
     virtual void inputHandler(
-      int index,
+      size_t index,
       int inputValue);
 
   private:
